+++ deepinv/tests/test_datasets.py
@@ -3,7 +3,7 @@ import shutil
 import PIL
 import pytest
 
-from deepinv.datasets import DIV2K, Urban100HR
+from deepinv.datasets import DIV2K, Urban100HR, Set14HR
 
 
 @pytest.fixture
@@ -11,8 +11,8 @@ def download_div2k():
     """Downloads dataset for tests and removes it after test executions."""
     tmp_data_dir = "DIV2K"
 
-    # Download div2K dataset
-    DIV2K(tmp_data_dir, download=True)
+    # Download div2K raw dataset
+    DIV2K(tmp_data_dir, mode="val", download=True)
 
     # This will return control to the test function
     yield tmp_data_dir
@@ -22,15 +22,14 @@ def download_div2k():
 
 
 def test_load_div2k_dataset(download_div2k):
-    """Check that DIV2K/DIV2K_train_HR contains 800 items."""
-    dataset = DIV2K(download_div2k, download=False)
-    assert len(dataset) == 800
-
-
-def test_load_div2K_img(download_div2k):
-    """Check that DIV2K/DIV2K_train_HR contains images."""
-    dataset = DIV2K(download_div2k, download=False)
-    assert type(dataset[0]) == PIL.PngImagePlugin.PngImageFile
+    """Check that DIV2K/DIV2K_train_HR contains 800 PIL images."""
+    val_dataset = DIV2K(download_div2k, mode="val", download=False)
+    assert (
+        len(val_dataset) == 100
+    ), f"Val dataset should have been of len 100, instead got {len(val_dataset)}."
+    assert (
+        type(val_dataset[0]) == PIL.PngImagePlugin.PngImageFile
+    ), "Dataset image should have been a PIL image."
 
 
 @pytest.fixture
@@ -38,7 +37,7 @@ def download_Urban100():
     """Downloads dataset for tests and removes it after test executions."""
     tmp_data_dir = "Urban100"
 
-    # Download div2K dataset
+    # Download Urban100 raw dataset
     Urban100HR(tmp_data_dir, download=True)
 
     # This will return control to the test function
@@ -49,16 +48,37 @@ def download_Urban100():
 
 
 def test_load_urban100_dataset(download_Urban100):
-    """Check that dataset contains 800 items."""
+    """Check that dataset contains 100 PIL images."""
     dataset = Urban100HR(download_Urban100, download=False)
     assert (
         len(dataset) == 100
     ), f"Dataset should have been of len 100, instead got {len(dataset)}."
+    assert (
+        type(dataset[0]) == PIL.PngImagePlugin.PngImageFile
+    ), "Dataset image should have been a PIL image."
 
 
-def test_load_urban100_img(download_Urban100):
-    """Check that dataset contains images."""
-    dataset = Urban100HR(download_Urban100, download=False)
+@pytest.fixture
+def download_Set14():
+    """Downloads dataset for tests and removes it after test executions."""
+    tmp_data_dir = "Set14"
+
+    # Download Set14 raw dataset
+    Set14HR(tmp_data_dir, download=True)
+
+    # This will return control to the test function
+    yield tmp_data_dir
+
+    # After the test function complete, any code after the yield statement will run
+    shutil.rmtree(tmp_data_dir)
+
+
+def test_load_set14_dataset(download_Set14):
+    """Check that dataset contains 14 PIL images."""
+    dataset = Set14HR(download_Set14, download=False)
+    assert (
+        len(dataset) == 14
+    ), f"Dataset should have been of len 14, instead got {len(dataset)}."
     assert (
         type(dataset[0]) == PIL.PngImagePlugin.PngImageFile
     ), "Dataset image should have been a PIL image."