+++ deepinv/training/trainer.py
@@ -115,7 +115,7 @@ class Trainer:
     :param str save_path: Directory in which to save the trained model.
     :param str device: Device on which to run the training (e.g., 'cuda' or 'cpu').
     :param bool verbose: Output training progress information in the console.
-    :param bool progress_bar: Show a progress bar during training.
+    :param bool show_progress_bar: Show a progress bar during training.
     :param bool plot_images: Plots reconstructions every ``ckp_interval`` epochs.
     :param bool wandb_vis: Use Weights & Biases visualization, see https://wandb.ai/ for more details.
     :param dict wandb_setup: Dictionary with the setup for wandb, see https://docs.wandb.ai/quickstart for more details.
@@ -171,14 +171,16 @@ class Trainer:
         if self.eval_dataloader and type(self.eval_dataloader) is not list:
             self.eval_dataloader = [self.eval_dataloader]
 
-        self.save_path = Path(self.save_path)
+        self.save_path = Path(self.save_path) if self.save_path else None
 
         self.eval_metrics_history = {}
         self.G = len(self.train_dataloader)
 
         if (
-            self.wandb_setup is not None or self.wandb_setup != {}
-        ) and not self.wandb_vis:
+            self.wandb_setup != {}
+            and self.wandb_setup is not None
+            and not self.wandb_vis
+        ):
             warnings.warn(
                 "wandb_vis is False but wandb_setup is provided. Wandb visualization deactivated (wandb_vis=False)."
             )
@@ -187,6 +189,18 @@ class Trainer:
             warnings.warn(
                 "Physics generator is provided but online_measurements is False. Physics generator will not be used."
             )
+
+        self.epoch_start = 0
+        if self.ckpt_pretrained is not None:
+            checkpoint = torch.load(self.ckpt_pretrained)
+            self.model.load_state_dict(checkpoint["state_dict"])
+            if "optimizer" in checkpoint:
+                self.optimizer.load_state_dict(checkpoint["optimizer"])
+            if "wandb_id" in checkpoint:
+                self.wandb_setup["id"] = checkpoint["wandb_id"]
+                self.wandb_setup["resume"] = "allow"
+            if "epoch" in checkpoint:
+                self.epoch_start = checkpoint["epoch"]
 
         # wandb initialization
         if self.wandb_vis:
@@ -225,7 +239,9 @@ class Trainer:
         if self.check_grad:
             self.check_grad_val = AverageMeter("Gradient norm", ":.2e")
 
-        self.save_path = f"{self.save_path}/{get_timestamp()}"
+        self.save_path = (
+            f"{self.save_path}/{get_timestamp()}" if self.save_path else None
+        )
 
         # count the overall training parameters
         params = sum(p.numel() for p in self.model.parameters() if p.requires_grad)
@@ -236,15 +252,6 @@ class Trainer:
             self.physics = [self.physics]
 
         self.loss_history = []
-
-        self.epoch_start = 0
-        if self.ckpt_pretrained is not None:
-            checkpoint = torch.load(self.ckpt_pretrained)
-            self.model.load_state_dict(checkpoint["state_dict"])
-            if "optimizer" in checkpoint:
-                self.optimizer.load_state_dict(checkpoint["optimizer"])
-            if "epoch" in checkpoint:
-                self.epoch_start = checkpoint["epoch"]
 
     def prepare_images(self, physics_cur, x, y, x_net):
         r"""
@@ -586,7 +593,7 @@ class Trainer:
                 log_dict_post_epoch["step"] = epoch
                 wandb.log(log_dict_post_epoch)
 
-    def save_model(self, epoch, eval_metrics=None):
+    def save_model(self, epoch, eval_metrics=None, state={}):
         r"""
         Save the model.
 
@@ -594,10 +601,15 @@ class Trainer:
 
         :param int epoch: Current epoch.
         :param None, float eval_metrics: Evaluation metrics across epochs.
+        :param dict state: custom objects to save with model
         """
+
+        if not self.save_path:
+            return
+
         if (epoch > 0 and epoch % self.ckp_interval == 0) or epoch + 1 == self.epochs:
             os.makedirs(str(self.save_path), exist_ok=True)
-            state = {
+            state = state | {
                 "epoch": epoch,
                 "state_dict": self.model.state_dict(),
                 "loss": self.loss_history,
@@ -605,6 +617,8 @@ class Trainer:
             }
             if eval_metrics is not None:
                 state["eval_metrics"] = eval_metrics
+            if self.wandb_vis:
+                state["wandb_id"] = wandb.run.id
             torch.save(
                 state,
                 os.path.join(
@@ -627,6 +641,20 @@ class Trainer:
         self.setup_train()
 
         for epoch in range(self.epoch_start, self.epochs):
+
+            # clean metrics
+            self.logs_total_loss_train.reset()
+            self.logs_total_loss_eval.reset()
+
+            for l in self.logs_losses_train:
+                l.reset()
+
+            for l in self.logs_metrics_train:
+                l.reset()
+
+            for l in self.logs_metrics_eval:
+                l.reset()
+
             ## Evaluation
             perform_eval = self.eval_dataloader and (
                 (epoch + 1) % self.eval_interval == 0 or epoch + 1 == self.epochs
@@ -674,13 +702,6 @@ class Trainer:
                     epoch, progress_bar, train=True, last_batch=(i == batches - 1)
                 )
 
-                if i < batches - 1:
-                    # clean meters
-                    self.logs_total_loss_train.reset()
-                    self.logs_total_loss_eval.reset()
-                    for l in self.logs_losses_train:
-                        l.reset()
-
             self.loss_history.append(self.logs_total_loss_train.avg)
 
             if self.scheduler:
@@ -711,8 +732,8 @@ class Trainer:
             online_measurements=self.online_measurements,
             plot_images=self.plot_images,
             plot_metrics=self.plot_metrics,
+            save_folder=self.save_path + "/test" if self.save_path else None,
             physics_generator=self.physics_generator,
-            save_folder=self.save_path + "/test",
             metrics=self.metrics,
             device=self.device,
             verbose=self.verbose,
@@ -732,7 +753,7 @@ def train(
     **kwargs,
 ):
     """
-    Alias function for training a model using :class:`deepinv.training_utils.Trainer` class.
+    Alias function for training a model using :class:`deepinv.Trainer` class.
 
     This function creates a Trainer instance and returns the trained model.